+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostSmartTimerTrigger/Example.kt
@@ -1,17 +1,4 @@
 package com.github.insanusmokrassar.AutoPostSmartTimerTrigger
 
-import com.github.insanusmokrassar.AutoPostSmartTimerTrigger.utils.near
-import com.github.insanusmokrassar.IObjectKRealisations.load
-import com.github.insanusmokrassar.IObjectKRealisations.toObject
-
 fun main(args: Array<String>) {
-    load(args[0]).also {
-        println(it)
-    }
-        .toObject(SmartTimerConfig::class.java).also {
-            println(
-                it.triggerTimes.joinToString("\n")
-            )
-            println("Near: ${it.triggerTimes.near()}")
-        }
 }