+++ CHANGELOG.md
@@ -32,3 +32,4 @@
 * Update telegram bot API library
 * Change version of build
 * Now plugin receive direct items
+* `SmartTimerConfigTimeItem` constructor fields are optional