+++ CHANGELOG.md
@@ -25,3 +25,11 @@
 ## 0.11
 
 * AutoPostTelegramBot `0.4.4` -> `0.4.5`
+
+## 1.0.0
+
+* Update kotlin version
+* Update telegram bot API library
+* Change version of build
+* Now plugin receive direct items
+* `SmartTimerConfigTimeItem` constructor fields are optional